Updates to math lib.

+++ lib/include/FenrirMath/type_vec3.hpp
@@ -37,8 +37,8 @@ namespace fenm {
 		type_vec3<_T>& operator -=(const type_vec3<_T>& v);
 		type_vec3<_T> operator *(const type_vec3<_T>& vec);
 		type_vec3<_T>& operator *=(const type_vec3<_T>& vec);
-		type_vec3<_T> operator *(const long& constant);
-		type_vec3<_T>& operator *=(const long& constant);
+		type_vec3<_T> operator *(const long long& constant);
+		type_vec3<_T>& operator *=(const long long& constant);
 		type_vec3<_T> operator *(const double& constant);
 		type_vec3<_T>& operator *=(const double& constant);
 		bool operator ==(const type_vec3<_T>& vec);
